+++ src/resources/views/permissions/include/_list.blade.php
@@ -1,26 +1,33 @@
 <div class="card">
     <div class="card-header">
-        <a href="{{ route('permissions.create') }}" class="btn btn-success btn-sm">
-            <i class="fas fa-plus"></i>
-            {{ trans('packages/core::common.create') }}
-        </a>
-        <button class="btn btn-sm bg-danger delete-all d-none ml-2" title="Delete" role="button" data-url="{{ route('permissions.deleteAll') }}">
-            <i class="fas fa-trash"></i>
-            {{ trans('packages/core::common.delete') }}
-        </button>
+        @can('create permissions')
+            <a href="{{ route('permissions.create') }}" class="btn btn-success btn-sm">
+                <i class="fas fa-plus"></i>
+                {{ trans('packages/core::common.create') }}
+            </a>
+        @endcan
+        @can('delete permissions')
+            <button class="btn btn-sm bg-danger delete-all d-none ml-2" title="Delete" role="button" data-url="{{ route('permissions.deleteAll') }}">
+                <i class="fas fa-trash"></i>
+                {{ trans('packages/core::common.delete') }}
+            </button>
+        @endcan
     </div>
 
     <div class="card-body">
         <table class="mt-3 table table-bordered table-hover table-striped" id="dataTable">
         <thead>
             <tr>
-                <th width="3%" class="text-center">
-                    <label class="user-checkbox-label">
-                        <input class="input-check-all" type="checkbox">
-                    </label>
-                </th>
-                <th width="20%">{{ trans('packages/core::common.name') }}</th>
-                <th width="45%">{{ trans('packages/core::common.role_permission.roles.title') }}</th>
+                @can('delete permissions')
+                    <th width="3%" class="text-center">
+                        <label class="user-checkbox-label">
+                            <input class="input-check-all" type="checkbox">
+                        </label>
+                    </th>
+                @endcan
+                <th width="15%">{{ trans('packages/core::common.name') }}</th>
+                <th width="15%">{{ trans('packages/core::common.role_permission.group_name') }}</th>
+                <th width="35%">{{ trans('packages/core::common.role_permission.roles.title') }}</th>
                 <th width="10%">{{ trans('packages/core::common.role_permission.guard_name') }}</th>
                 <th width="12%" class="text-center">{{ trans('packages/core::common.created_at') }}</th>
                 <th width="10%" class="text-center">{{ trans('packages/core::common.operations') }}</th>
@@ -29,14 +36,21 @@
         <tbody>
         @foreach($permissions as $item)
             <tr>
-                <td class="text-center">
-                    <label class="user-checkbox-label">
-                        <input class="checkboxes" type="checkbox" value="{{ $item->id }}">
-                    </label>
-                </td>
+                @can('delete permissions')
+                    <td class="text-center">
+                        <label class="user-checkbox-label">
+                            <input class="checkboxes" type="checkbox" value="{{ $item->id }}">
+                        </label>
+                    </td>
+                @endcan
                 <td>
-                    <a href="{{ route('permissions.edit', $item->id) }}">{{ $item->name }}</a>
+                    @can('edit permissions')
+                        <a href="{{ route('permissions.edit', $item->id) }}">{{ $item->name }}</a>
+                    @else
+                        {{ $item->name }}
+                    @endcan
                 </td>
+                <td>{{ $item->group_name }}</td>
                 <td>
                     @if($item->roles->isNotEmpty())
                         @foreach($item->roles as $role)
@@ -44,15 +58,19 @@
                         @endforeach
                     @endif
                 </td>
-                <td>{{ $item->guard_name }}</td>
+                <td>{{ $item->group_name }}</td>
                 <td class="text-center">{{ $item->created_at }}</td>
                 <td class="text-center">
-                    <a class="btn btn-sm bg-info" href="{{ route('permissions.edit', $item->id) }}">
-                        <i class="fas fa-pencil-alt"></i>
-                    </a>
-                    <button type="button" class="btn btn-sm bg-danger btn-delete" data-url="{{ route('permissions.destroy', $item->id) }}">
-                        <i class="fas fa-trash"></i>
-                    </button>
+                    @can('edit permissions')
+                        <a class="btn btn-sm bg-info" href="{{ route('permissions.edit', $item->id) }}">
+                            <i class="fas fa-pencil-alt"></i>
+                        </a>
+                    @endcan
+                    @can('delete permissions')
+                        <button type="button" class="btn btn-sm bg-danger btn-delete" data-url="{{ route('permissions.destroy', $item->id) }}">
+                            <i class="fas fa-trash"></i>
+                        </button>
+                    @endcan
                 </td>
             </tr>
         @endforeach
